refactor: move logic of collapse mobile menu into header.tsx
